refactor(filter): Consolidate end-to-end tests
