+++ crates/snapbox/src/filter/test.rs
@@ -1,8 +1,245 @@
+use std::path::PathBuf;
+
 #[cfg(feature = "json")]
 use serde_json::json;
 
-#[cfg(feature = "json")]
 use super::*;
+use crate::prelude::*;
+
+#[test]
+fn str_normalize_redactions_empty() {
+    let input = "";
+    let pattern = "";
+    let expected = "";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_literals_match() {
+    let input = "Hello\nWorld";
+    let pattern = "Hello\nWorld";
+    let expected = "Hello\nWorld";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_pattern_shorter() {
+    let input = "Hello\nWorld";
+    let pattern = "Hello\n";
+    let expected = "Hello\nWorld";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_input_shorter() {
+    let input = "Hello\n";
+    let pattern = "Hello\nWorld";
+    let expected = "Hello\n";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_all_different() {
+    let input = "Hello\nWorld";
+    let pattern = "Goodbye\nMoon";
+    let expected = "Hello\nWorld";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_middles_diverge() {
+    let input = "Hello\nWorld\nGoodbye";
+    let pattern = "Hello\nMoon\nGoodbye";
+    let expected = "Hello\nWorld\nGoodbye";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_elide_delimited_with_sub() {
+    let input = "Hello World\nHow are you?\nGoodbye World";
+    let pattern = "Hello [..]\n...\nGoodbye [..]";
+    let expected = "Hello [..]\n...\nGoodbye [..]";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_leading_elide() {
+    let input = "Hello\nWorld\nGoodbye";
+    let pattern = "...\nGoodbye";
+    let expected = "...\nGoodbye";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_trailing_elide() {
+    let input = "Hello\nWorld\nGoodbye";
+    let pattern = "Hello\n...";
+    let expected = "Hello\n...";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_middle_elide() {
+    let input = "Hello\nWorld\nGoodbye";
+    let pattern = "Hello\n...\nGoodbye";
+    let expected = "Hello\n...\nGoodbye";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_post_elide_diverge() {
+    let input = "Hello\nSun\nAnd\nWorld";
+    let pattern = "Hello\n...\nMoon";
+    let expected = "Hello\nSun\nAnd\nWorld";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_post_diverge_elide() {
+    let input = "Hello\nWorld\nGoodbye\nSir";
+    let pattern = "Hello\nMoon\nGoodbye\n...";
+    let expected = "Hello\nWorld\nGoodbye\nSir";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_inline_elide() {
+    let input = "Hello\nWorld\nGoodbye\nSir";
+    let pattern = "Hello\nW[..]d\nGoodbye\nSir";
+    let expected = "Hello\nW[..]d\nGoodbye\nSir";
+    let actual = NormalizeToExpected::new()
+        .redact()
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, expected.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_user_literal() {
+    let input = "Hello world!";
+    let pattern = "Hello [OBJECT]!";
+    let mut sub = Redactions::new();
+    sub.insert("[OBJECT]", "world").unwrap();
+    let actual = NormalizeToExpected::new()
+        .redact_with(&sub)
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, pattern.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_user_path() {
+    let input = "input: /home/epage";
+    let pattern = "input: [HOME]";
+    let mut sub = Redactions::new();
+    let sep = std::path::MAIN_SEPARATOR.to_string();
+    let redacted = PathBuf::from(sep).join("home").join("epage");
+    sub.insert("[HOME]", redacted).unwrap();
+    let actual = NormalizeToExpected::new()
+        .redact_with(&sub)
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, pattern.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_user_overlapping_path() {
+    let input = "\
+a: /home/epage
+b: /home/epage/snapbox";
+    let pattern = "\
+a: [A]
+b: [B]";
+    let mut sub = Redactions::new();
+    let sep = std::path::MAIN_SEPARATOR.to_string();
+    let redacted = PathBuf::from(&sep).join("home").join("epage");
+    sub.insert("[A]", redacted).unwrap();
+    let redacted = PathBuf::from(sep)
+        .join("home")
+        .join("epage")
+        .join("snapbox");
+    sub.insert("[B]", redacted).unwrap();
+    let actual = NormalizeToExpected::new()
+        .redact_with(&sub)
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, pattern.into_data());
+}
+
+#[test]
+fn str_normalize_redactions_user_disabled() {
+    let input = "cargo";
+    let pattern = "cargo[EXE]";
+    let mut sub = Redactions::new();
+    sub.insert("[EXE]", "").unwrap();
+    let actual = NormalizeToExpected::new()
+        .redact_with(&sub)
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, pattern.into_data());
+}
+
+#[test]
+#[cfg(feature = "regex")]
+fn str_normalize_redactions_user_regex_unnamed() {
+    let input = "Hello world!";
+    let pattern = "Hello [OBJECT]!";
+    let mut sub = Redactions::new();
+    sub.insert("[OBJECT]", regex::Regex::new("world").unwrap())
+        .unwrap();
+    let actual = NormalizeToExpected::new()
+        .redact_with(&sub)
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, pattern.into_data());
+}
+
+#[test]
+#[cfg(feature = "regex")]
+fn str_normalize_redactions_user_regex_named() {
+    let input = "Hello world!";
+    let pattern = "Hello [OBJECT]!";
+    let mut sub = Redactions::new();
+    sub.insert(
+        "[OBJECT]",
+        regex::Regex::new("(?<redacted>world)!").unwrap(),
+    )
+    .unwrap();
+    let actual = NormalizeToExpected::new()
+        .redact_with(&sub)
+        .normalize(input.into(), &pattern.into());
+    assert_eq!(actual, pattern.into_data());
+}
 
 // Tests for normalization on json
 #[test]
